% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm.R
\name{btm}
\alias{btm}
\title{Bayesian Biterm Topic Model with Stan}
\usage{
btm(biterms, k = 5, alpha = 50/k, beta = 0.01, ...)
}
\arguments{
\item{k}{Number of topics.}

\item{alpha}{Hyperparameter vector for the Dirirchlet topic prior.}

\item{beta}{Hyperparameter vector for the Dirirchlet term priors.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian Biterm Topic Model with Stan
}
