% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_biterms.R
\name{id_terms}
\alias{id_terms}
\title{Convert terms to integers}
\usage{
id_terms(x)
}
\arguments{
\item{x}{A data.frame of biterms, with at least two columns: term1 and term2.}
}
\value{
The same data.frame, with terms converted to integers for Stan.
}
\description{
Convert terms to integers
}
